email: Move `PossibleTyposquatEmail` struct to its corresponding module
