+++ src/worker/jobs/typosquat.rs
@@ -5,7 +5,7 @@ use crates_io_worker::BackgroundJob;
 use diesel::PgConnection;
 use typomania::Package;
 
-use crate::email::PossibleTyposquatEmail;
+use crate::email::Email;
 use crate::tasks::spawn_blocking;
 use crate::{
     typosquat::{Cache, Crate},
@@ -80,6 +80,41 @@ fn check(
     }
 
     Ok(())
+}
+
+#[derive(Debug, Clone)]
+struct PossibleTyposquatEmail<'a> {
+    domain: &'a str,
+    crate_name: &'a str,
+    squats: &'a [typomania::checks::Squat],
+}
+
+impl Email for PossibleTyposquatEmail<'_> {
+    const SUBJECT: &'static str = "Possible typosquatting in new crate";
+
+    fn body(&self) -> String {
+        let squats = self
+            .squats
+            .iter()
+            .map(|squat| {
+                let domain = self.domain;
+                let crate_name = squat.package();
+                format!("- {squat} (https://{domain}/crates/{crate_name})\n")
+            })
+            .collect::<Vec<_>>()
+            .join("");
+
+        format!(
+            "New crate {crate_name} may be typosquatting one or more other crates.\n
+Visit https://{domain}/crates/{crate_name} to see the offending crate.\n
+\n
+Specific squat checks that triggered:\n
+\n
+{squats}",
+            domain = self.domain,
+            crate_name = self.crate_name,
+        )
+    }
 }
 
 #[cfg(test)]
